Updating Pyside2 legacy code to be compatible with Pyside6
PySide6.QtGui.QFont.setWeight() Isn't accepting integer values for font weight, so instances of this code were replaced with PySide6.QtGui.QFont.setLegacyWeight(), which preserves the old system of setting font weight by integer values. Also, matplotlib.backends.backend_qt5agg does not support Pyside6, so this line of code was updated to the new Pyside6 compatible matplotlib.backends.backend_qtagg. Finally, app.exec_() is soon to be deprecated, so it was replaced with the up-to-date form app.exec().